Projects: Fix panic caused by newly added enum (#443)
Recently following enums where added to project graphql query
https://docs.github.com/en/graphql/overview/changelog#schema-changes-for-2025-02-01-1:

- Enum value 'PARENT_ISSUEwas added to enumProjectV2FieldType'
- Enum value 'SUB_ISSUES_PROGRESSwas added to enumProjectV2FieldType'


As we didn't have processing for these, and they weren't in the
`fieldTypes` or `exclude` list `fieldType :=
fieldTypes[f.Common.DataType]` returned `nil` which caused panic in
`field := data.NewField(f.Common.Name, nil, fieldType)` as fieldType
can't be nil.

To fix this, we have added these new enums to `exclude` so we can add
processing for them later. Moreover, we have added a logic to check if
`fieldType == nil` and in that case `continue`, but create a log so we
might be more aware of this new enum.


More in
https://raintank-corp.slack.com/archives/C05K271Q8F5/p1742559654567789?thread_ts=1742486345.609529&cid=C05K271Q8F5